Remove all the old input components (I could find)
