--- inst/ui-editor-react/src/components/Inputs/InputSections.module.css
@@ -1,19 +1,0 @@
-.wrappedSection {
-  display: flex;
-  border: none;
-  flex-wrap: wrap;
-  justify-content: space-between;
-  margin-block-end: var(--vertical-spacing);
-}
-
-.sectionContainer {
-  margin-block-start: 25px;
-}
-
-.wrappedSection > * {
-  margin-bottom: 0;
-}
-
-.wrappedSection input {
-  max-width: unset;
-}
