--- inst/ui-editor-react/src/components/Inputs/Inputs.module.css
@@ -1,85 +1,0 @@
-/* Styles common to multiple input components */
-input {
-  padding: var(--input-vertical-padding) var(--input-horizontal-padding);
-  border: 1px solid var(--light-grey);
-  border-radius: var(--corner-radius);
-}
-
-.container {
-  --vertical-pad: calc(var(--vertical-spacing) * 1.5);
-  margin-top: var(--vertical-pad);
-  margin-bottom: var(--vertical-pad);
-  max-width: 100%;
-  min-width: 0;
-  position: relative;
-  isolation: isolate;
-  display: block;
-}
-
-.container[data-width-setting="fit"] {
-  width: fit-content;
-}
-.container[data-width-setting="full"] {
-  width: 100%;
-}
-
-.label {
-  display: block;
-  padding-bottom: 3px;
-  width: auto;
-  display: flex;
-  justify-items: flex-start;
-  align-items: center;
-  gap: 8px;
-  text-transform: capitalize;
-}
-
-.label input[type="checkbox"] {
-  width: fit-content;
-  /* By shifting the box down a tiny bit we balance the vertical alignment */
-  margin-bottom: -2px;
-}
-
-.container input {
-  max-width: 100%;
-  /* width: 100%; */
-}
-
-input:disabled,
-select:disabled {
-  outline: var(--disabled-outline);
-  opacity: 0.25;
-  cursor: not-allowed;
-}
-
-.container[data-disabled="true"] {
-  color: var(--disabled-color);
-}
-
-.mainInput {
-  position: relative;
-}
-
-.container[data-disabled="true"] .mainInput::before {
-  content: "";
-  position: absolute;
-  inset: 0;
-  border-radius: var(--corner-radius);
-  border: var(--disabled-outline);
-  background-color: var(--rstudio-white);
-  z-index: 1;
-  pointer-events: none;
-}
-.container[data-disabled="true"] .mainInput::after {
-  content: "Default";
-  position: absolute;
-  left: 50%;
-  top: 50%;
-  transform: translate(-50%, -50%);
-  pointer-events: none;
-  z-index: 2;
-}
-
-.mainInput > input[type="checkbox"] {
-  width: fit-content;
-}
