+++ inst/ui-editor-react/src/components/Inputs/ListInput/NamedListInput.tsx
@@ -1,20 +1,18 @@
 import React from "react";
 
-import CategoryDivider from "components/CategoryDivider";
 import { Trash } from "components/Icons";
 import { FaPlus } from "react-icons/fa";
 import { MdDragHandle } from "react-icons/md";
 import { ReactSortable } from "react-sortablejs";
 import { sameObject } from "utils/equalityCheckers";
 
-import type { InputWidgetCommonPropsOld } from "..";
 import Button from "../Button/Button";
-import type { OnChangeCallback } from "../SettingsUpdateContext";
-import { useOnChange } from "../SettingsUpdateContext";
+import type { InputComponentProps } from "../SettingsFormBuilder/inputFieldTypes";
+import { makeLabelId } from "../SettingsFormBuilder/inputFieldTypes";
 
 import classes from "./styles.module.css";
 
-export type ItemType = {
+type ItemType = {
   id: number;
   key: string;
   value: string;
@@ -33,23 +31,107 @@ export function isNamedList(x: any): x is NamedList {
   return true;
 }
 
-export default function NamedListInput({
-  name,
-  label,
+type NewItemValue = { key: string; value: string };
+
+export function NamedListInput({
+  id,
   value,
   onChange,
-  optional = false,
   newItemValue = { key: "myKey", value: "myValue" },
-}: InputWidgetCommonPropsOld<NamedList> & {
-  newItemValue?: { key: string; value: string };
+}: InputComponentProps<
+  NamedList,
+  {
+    newItemValue?: NewItemValue;
+  }
+>) {
+  const { state, setState, addItem, deleteItem } = useListState({
+    value,
+    onChange,
+    newItemValue,
+  });
+
+  return (
+    <div
+      className={classes.list}
+      aria-labelledby={makeLabelId(id)}
+      aria-label="Reorderable list input"
+    >
+      <div
+        className={classes.item + " " + classes.header}
+        aria-label="Columns field labels"
+      >
+        <span className={classes.keyField}>Key</span>
+        <span className={classes.valueField}>Value</span>
+      </div>
+      <ReactSortable
+        list={state}
+        setList={setState}
+        handle={`.${classes.dragHandle}`}
+      >
+        {state.map((item, i) => (
+          <div className={classes.item} key={item.id}>
+            <div className={classes.dragHandle} title="Reorder list">
+              <MdDragHandle />
+            </div>
+            <input
+              className={classes.keyField}
+              type="text"
+              value={item.key}
+              onChange={(e) => {
+                const newList = [...state];
+                newList[i] = { ...item, key: e.target.value };
+                setState(newList);
+              }}
+            />
+            <span className={classes.separator}>:</span>
+            <input
+              className={classes.valueField}
+              type="text"
+              value={item.value}
+              onChange={(e) => {
+                const newList = [...state];
+                newList[i] = { ...item, value: e.target.value };
+                setState(newList);
+              }}
+            />
+            <Button
+              className={classes.deleteButton}
+              onClick={() => deleteItem(item.id)}
+              variant={["icon", "transparent"]}
+              title={`Delete ${item.value}`}
+            >
+              <Trash />
+            </Button>
+          </div>
+        ))}
+      </ReactSortable>
+      <Button
+        className={classes.addItemButton}
+        onClick={() => addItem()}
+        variant={["icon", "transparent"]}
+        title="Add new item"
+        aria-label="Add new item to list"
+      >
+        <FaPlus />
+      </Button>
+    </div>
+  );
+}
+
+function useListState({
+  value,
+  onChange,
+  newItemValue,
+}: {
+  value: NamedList;
+  onChange: (x: NamedList) => void;
+  newItemValue: NewItemValue;
 }) {
   const [state, setState] = React.useState<ItemType[]>(
     value !== undefined
       ? Object.keys(value).map((key, i) => ({ id: i, key, value: value[key] }))
       : []
   );
-
-  const onNewValue = useOnChange(onChange as OnChangeCallback);
 
   React.useEffect(() => {
     const newList = simplifyToChoices(state);
@@ -59,8 +141,8 @@ export default function NamedListInput({
       return;
     }
 
-    onNewValue({ name, value: newList });
-  }, [name, onNewValue, state, value]);
+    onChange(newList);
+  }, [onChange, state, value]);
 
   const deleteItem = React.useCallback((itemId: number) => {
     setState((list) => list.filter(({ id }) => id !== itemId));
@@ -77,70 +159,14 @@ export default function NamedListInput({
     );
   }, [newItemValue]);
 
-  return (
-    <div className={classes.container}>
-      <CategoryDivider category={name ?? label} />
-      <div className={classes.list}>
-        <div className={classes.item + " " + classes.header}>
-          <span className={classes.keyField}>Key</span>
-          <span className={classes.valueField}>Value</span>
-        </div>
-        <ReactSortable
-          list={state}
-          setList={setState}
-          handle={`.${classes.dragHandle}`}
-        >
-          {state.map((item, i) => (
-            <div className={classes.item} key={item.id}>
-              <div className={classes.dragHandle} title="Reorder list">
-                <MdDragHandle />
-              </div>
-              <input
-                className={classes.keyField}
-                type="text"
-                value={item.key}
-                onChange={(e) => {
-                  const newList = [...state];
-                  newList[i] = { ...item, key: e.target.value };
-                  setState(newList);
-                }}
-              />
-              <span className={classes.separator}>:</span>
-              <input
-                className={classes.valueField}
-                type="text"
-                value={item.value}
-                onChange={(e) => {
-                  const newList = [...state];
-                  newList[i] = { ...item, value: e.target.value };
-                  setState(newList);
-                }}
-              />
-              <Button
-                className={classes.deleteButton}
-                onClick={() => deleteItem(item.id)}
-                variant={["icon", "transparent"]}
-                title={`Delete ${item.value}`}
-              >
-                <Trash />
-              </Button>
-            </div>
-          ))}
-        </ReactSortable>
-        <Button
-          className={classes.addItemButton}
-          onClick={() => addItem()}
-          variant={["icon", "transparent"]}
-          title="Add new item"
-        >
-          <FaPlus />
-        </Button>
-      </div>
-    </div>
-  );
+  return {
+    state,
+    setState,
+    deleteItem,
+    addItem,
+  };
 }
-
-export function simplifyToChoices(arrayVersion: ItemType[]): NamedList {
+function simplifyToChoices(arrayVersion: ItemType[]): NamedList {
   const toReturn: NamedList = arrayVersion.reduce(
     (namedList, { key, value }) => {
       namedList[key] = value;
