--- inst/ui-editor-react/src/components/Inputs/OptionalInput/styles.module.css
@@ -1,5 +1,0 @@
-.container {
-  display: flex;
-  gap: 5px;
-  align-items: center;
-}
