--- inst/ui-editor-react/src/components/Inputs/TextInput/TextInput.module.css
@@ -1,19 +1,0 @@
-.container {
-  display: flex;
-  gap: 0.5rem;
-  flex-wrap: wrap;
-  align-items: center;
-  margin-bottom: 1rem;
-  width: 100%;
-}
-
-.label {
-  font-style: italic;
-  width: auto;
-}
-
-.input {
-  border: 1px solid var(--light-grey);
-  border-radius: var(--corner-radius);
-  width: 100%;
-}
